+++ src/helpers/output-info.ts
@@ -11,10 +11,10 @@ import {
 } from 'src/constants/component';
 
 import {Logger} from './logger';
-import {PasCalCase} from './utils';
+import {PasCalCase, fillAnsiLength, strip} from './utils';
 
 // eslint-disable-next-line no-control-regex
-const colorMatchRegex = /\u001b\[[\d;]+m/g;
+export const colorMatchRegex = /\u001b\[[\d;]+m/g;
 
 const rounded = boxRound.round;
 const space = '   ';
@@ -46,11 +46,13 @@ export function outputComponents({
     return;
   }
 
-  const componentKeyLengthMap: Record<keyof NextUIComponent, number> = {
+  const componentKeyLengthMap: Record<keyof NextUIComponent | 'originVersion', number> = {
     description: 0,
     docs: 0,
     name: 0,
+    originVersion: 0,
     package: 0,
+    peerDependencies: 0,
     status: 0,
     style: 0,
     version: 0
@@ -58,13 +60,29 @@ export function outputComponents({
 
   for (const component of components) {
     for (const key in component) {
-      // Align the length of the version
-      componentKeyLengthMap[key] = Math.max(
-        componentKeyLengthMap[key],
-        key === 'version'
-          ? Math.max(String(component[key]).length, 'version'.length)
-          : String(component[key]).length
-      );
+      const str = String(component[key]);
+
+      if (key === 'version') {
+        const newVersion = str.match(/new:\s([\d.]+)/)?.[1];
+        const currentVersion = str.match(/([\d.]+)\snew:/)?.[1];
+
+        const value = strip(generateComponentOutputVersion(currentVersion!, newVersion!));
+
+        // Align the length of the version
+        componentKeyLengthMap[key] = Math.max(
+          componentKeyLengthMap[key],
+          Math.max(value.length, 'version'.length)
+        );
+        // Record origin version length
+        componentKeyLengthMap.originVersion = Math.max(
+          componentKeyLengthMap.originVersion,
+          currentVersion!.length
+        );
+
+        continue;
+      }
+
+      componentKeyLengthMap[key] = Math.max(componentKeyLengthMap[key], str.length);
     }
   }
 
@@ -72,7 +90,7 @@ export function outputComponents({
     let outputData = padStart;
 
     for (const key of orderNextUIComponentKeys) {
-      let value = component[key].padEnd(componentKeyLengthMap[key]);
+      let value = fillAnsiLength(component[key], componentKeyLengthMap[key]);
 
       /** ======================== Replace version to new version ======================== */
       if (commandName !== 'list' && key === 'version') {
@@ -81,17 +99,18 @@ export function outputComponents({
         const newVersion = value.match(/new:\s([\d.]+)/)?.[1];
 
         if (currentVersion === newVersion) {
-          value = value.replace(/\snew:\s([\d.]+)/, '');
-          value = `${value} 🚀latest`.padEnd(componentKeyLengthMap[key]);
+          value = value.replace(/\snew:\s[\d.]+(\s+)?/, '');
+          value = fillAnsiLength(
+            `${fillAnsiLength(value, componentKeyLengthMap.originVersion)} 🚀latest`,
+            componentKeyLengthMap[key]
+          );
           value = value.replace('latest', chalk.magentaBright.underline('latest'));
         } else if (newVersion) {
-          value = `${chalk.white(`${currentVersion} ->`)} ${chalk.yellowBright(
-            `${newVersion} (new)`
-          )}`;
-
-          componentKeyLengthMap[key] = Math.max(
-            // eslint-disable-next-line no-control-regex
-            value.replace(/(\u001b\[\d+m)/g, '').length,
+          value = fillAnsiLength(
+            generateComponentOutputVersion(
+              fillAnsiLength(currentVersion!, componentKeyLengthMap.originVersion),
+              newVersion
+            ),
             componentKeyLengthMap[key]
           );
         }
@@ -178,8 +197,10 @@ export function outputInfo() {
  * @param title
  * @param borderStyle
  * @param padding
+ * @param align
  */
 export function outputBox({
+  align = 'center',
   borderStyle = 'round',
   center = false,
   color,
@@ -195,6 +216,7 @@ export function outputBox({
   title?: string;
   borderStyle?: keyof typeof boxRound;
   padding?: number;
+  align?: 'left' | 'center' | 'right';
 }) {
   const rounded = boxRound[borderStyle];
   const mergedRounded = color
@@ -226,22 +248,30 @@ export function outputBox({
   // Update the titleHeaderLength
   titleHeaderLength = maxLength - titleLength;
 
-  const boxHeaderContent = title
-    ? `${rounded.horizontal
-        .padEnd(Math.floor(titleHeaderLength / 2) - 1, rounded.horizontal)
-        .replaceAll(rounded.horizontal, mergedRounded.horizontal)} ${title} ${rounded.horizontal
-        .padEnd(Math.ceil(titleHeaderLength / 2) - 1, rounded.horizontal)
-        .replaceAll(rounded.horizontal, mergedRounded.horizontal)}`
-    : rounded.horizontal
-        .padEnd(maxLength, rounded.horizontal)
-        .replaceAll(rounded.horizontal, mergedRounded.horizontal);
+  const boxHeaderContent = (() => {
+    if (title) {
+      if (align === 'center') {
+        const spaceFir = Math.floor(titleHeaderLength / 2) - 1;
+        const spaceSec = Math.ceil(titleHeaderLength / 2) - 1;
+
+        const padFir = spaceFir > 0 ? mergedRounded.horizontal.repeat(spaceFir) : '';
+        const padSec = spaceSec > 0 ? mergedRounded.horizontal.repeat(spaceSec) : '';
+
+        return `${padFir} ${title} ${padSec}`;
+      } else if (align === 'left') {
+        return ` ${title} ${mergedRounded.horizontal.repeat(titleHeaderLength - 2)}`;
+      } else {
+        return `${mergedRounded.horizontal.repeat(titleHeaderLength - 2)} ${title} `;
+      }
+    }
+
+    return mergedRounded.horizontal.repeat(maxLength);
+  })();
 
   const boxHeader = mergedRounded.topLeft + boxHeaderContent + mergedRounded.topRight;
   const boxFooter =
     mergedRounded.bottomLeft +
-    rounded.horizontal
-      .padEnd(maxLength, rounded.horizontal)
-      .replaceAll(rounded.horizontal, mergedRounded.horizontal) +
+    mergedRounded.horizontal.repeat(maxLength) +
     mergedRounded.bottomRight;
 
   let boxContent = contentArr.reduce((acc, cur) => {
@@ -259,17 +289,27 @@ export function outputBox({
     // Over 2 cause one vertical line == 2 spaces
     // paddingLength = Math.floor(Math.max(paddingLength, spaceFir, spaceSec) / 2);
 
-    mergedPadding
-      ? acc.push(
-          `${mergedRounded.vertical}${spaceLength ? `${padFir}${cur}${padSec}` : cur}${
-            mergedRounded.vertical
-          }`
-        )
-      : acc.push(
-          `${mergedRounded.vertical}${spaceLength > 0 ? `${cur}${pad}` : cur}${
-            mergedRounded.vertical
-          }`
-        );
+    if (center) {
+      acc.push(
+        `${mergedRounded.vertical}${spaceLength ? `${padFir}${cur}${padSec}` : cur}${
+          mergedRounded.vertical
+        }`
+      );
+    } else if (padding) {
+      const endLen = spaceLength - paddingLength * 2;
+
+      acc.push(
+        `${mergedRounded.vertical}${' '.repeat(paddingLength * 2)}${cur}${' '.repeat(endLen)}${
+          mergedRounded.vertical
+        }`
+      );
+    } else {
+      acc.push(
+        `${mergedRounded.vertical}${spaceLength > 0 ? `${cur}${pad}` : cur}${
+          mergedRounded.vertical
+        }`
+      );
+    }
 
     return acc;
   }, [] as string[]);
@@ -289,4 +329,8 @@ export function outputBox({
   log && Logger.log(boxContent.join('\n'));
 
   return boxContent.join('\n');
+}
+
+function generateComponentOutputVersion(currentVersion: string, newVersion: string) {
+  return `${chalk.white(`${currentVersion} ->`)} ${chalk.yellowBright(`${newVersion} (new)`)}`;
 }