+++ src/prompts/index.ts
@@ -10,13 +10,13 @@ const defaultPromptOptions: prompts.Options = {
   }
 };
 
-export async function getText(message: string, initial: string) {
+export async function getText(message: string, initial?: string) {
   const result = await prompts(
     {
-      initial,
       message,
       name: 'value',
-      type: 'text'
+      type: 'text',
+      ...(initial ? {initial} : {})
     },
     defaultPromptOptions
   );
@@ -60,6 +60,22 @@ export async function getSelect(message: string, choices: prompts.Choice[]) {
       message,
       name: 'value',
       type: 'select',
+      ...(choices ? {choices} : {})
+    },
+    defaultPromptOptions
+  );
+
+  return result.value;
+}
+
+export async function getMultiselect(message: string, choices?: prompts.Choice[]) {
+  const result = await prompts(
+    {
+      hint: '- Space to select. Return to submit',
+      message,
+      min: 1,
+      name: 'value',
+      type: 'multiselect',
       ...(choices ? {choices} : {})
     },
     defaultPromptOptions